Clean up code by splitting up components.
