+++ client/src/pages/studentmanagement/student-overview-new/StudentList.tsx
@@ -1,64 +1,19 @@
-import { Box, Button, Paper, PaperProps, TextField, Typography } from '@material-ui/core';
-import { createStyles, makeStyles } from '@material-ui/core/styles';
+import { Box } from '@material-ui/core';
 import { FormikHelpers } from 'formik';
-import {
-  AccountSearch as SearchIcon,
-  AccountSwitch as ChangeTutorialIcon,
-  InformationOutline as InfoIcon,
-  Mail as MailIcon,
-} from 'mdi-material-ui';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { Link } from 'react-router-dom';
 import { FixedSizeList } from 'react-window';
-import {
-  ScheinCriteriaSummary,
-  ScheincriteriaSummaryByStudents,
-} from 'shared/model/ScheinCriteria';
+import { ScheincriteriaSummaryByStudents } from 'shared/model/ScheinCriteria';
 import { IStudentDTO } from 'shared/model/Student';
-import CustomSelect from '../../../components/CustomSelect';
 import StudentForm, {
   convertFormStateToDTO,
   StudentFormState,
 } from '../../../components/forms/StudentForm';
-import EntityListItemMenu from '../../../components/list-item-menu/EntityListItemMenu';
-import { ListItem } from '../../../components/list-item-menu/ListItemMenu';
-import StudentAvatar from '../../../components/student-icon/StudentAvatar';
 import { useDialog } from '../../../hooks/dialog-service/DialogService';
 import { Student } from '../../../model/Student';
 import { Team } from '../../../model/Team';
-import { ROUTES } from '../../../routes/Routing.routes';
-import ScheinStatusBox from '../../student-info/components/ScheinStatusBox';
+import StudentListRow, { SubtextType } from './components/StudentListRow';
+import StudentListTopBar from './components/StudentListTopBar';
 import { getFilteredStudents, StudentSortOption } from './StudentList.helpers';
-
-const useStyles = makeStyles((theme) =>
-  createStyles({
-    studentBar: {
-      display: 'flex',
-      alignItems: 'center',
-      padding: theme.spacing(2),
-      '&:hover': {
-        background: theme.palette.action.hover,
-      },
-    },
-    searchField: {
-      flex: 1,
-    },
-    sortSelect: {
-      marginLeft: theme.spacing(2),
-      minWidth: '20%',
-    },
-  })
-);
-
-type SubtextType = 'team' | 'tutorial';
-
-interface TopBarProps {
-  filterText: string;
-  onFilterTextChanged: (newText: string) => void;
-  sortOption: StudentSortOption;
-  onSortOptionChanged: (newOption: StudentSortOption) => void;
-  additionalTopBarItem?: React.ReactNode;
-}
 
 interface Props {
   students: Student[];
@@ -76,165 +31,7 @@ interface Props {
   tutorialId?: string;
 }
 
-function StudentListTopBar({
-  filterText,
-  onFilterTextChanged,
-  sortOption,
-  onSortOptionChanged,
-  additionalTopBarItem,
-}: TopBarProps): JSX.Element {
-  const classes = useStyles();
-
-  const handleSortOptionChanged = useCallback(
-    (e: React.ChangeEvent<{ name?: string; value: unknown }>) => {
-      if (typeof e.target.value !== 'string') {
-        return;
-      }
-
-      if (e.target.value === sortOption) {
-        return;
-      }
-
-      const selectedOption: StudentSortOption | undefined = Object.values(StudentSortOption).find(
-        (op) => op === e.target.value
-      );
-
-      if (!selectedOption) {
-        throw new Error('Selected filter option is not a valid one.');
-      }
-
-      onSortOptionChanged(selectedOption);
-    },
-    [sortOption, onSortOptionChanged]
-  );
-
-  return (
-    <Box display='flex'>
-      <TextField
-        variant='outlined'
-        label='Suche'
-        value={filterText}
-        onChange={(e) => onFilterTextChanged(e.target.value)}
-        className={classes.searchField}
-        InputProps={{
-          startAdornment: <SearchIcon color='disabled' />,
-        }}
-      />
-
-      <CustomSelect
-        label='Sortieren nach...'
-        emptyPlaceholder='Keine Sortieroptionen vorhanden.'
-        className={classes.sortSelect}
-        value={sortOption}
-        items={Object.values(StudentSortOption)}
-        itemToString={(option) => option}
-        itemToValue={(option) => option}
-        onChange={handleSortOptionChanged}
-      />
-
-      {additionalTopBarItem}
-    </Box>
-  );
-}
-
 const GUTTER_SIZE = 16;
-
-interface GetSubtextParams {
-  student: Student;
-  subTextType: SubtextType;
-}
-
-function getSubtext({ student, subTextType }: GetSubtextParams): string {
-  const { team, tutorial } = student ?? { team: undefined, tutorial: undefined };
-
-  switch (subTextType) {
-    case 'team':
-      return team ? `Team: #${team.teamNo.toString().padStart(2, '0')}` : 'Kein Team.';
-
-    case 'tutorial':
-      return tutorial ? `Tutorium: ${tutorial.slot}` : 'In keinem Tutorium.';
-  }
-}
-
-interface StudentListRowProps extends PaperProps {
-  student: Student;
-  scheinStatus: ScheinCriteriaSummary;
-  subTextType: SubtextType;
-  onEdit: (student: Student) => void;
-  onDelete: (student: Student) => void;
-  onChangeTutorial?: (student: Student) => void;
-  tutorialId?: string;
-}
-
-function StudentListRow({
-  student,
-  scheinStatus,
-  subTextType,
-  tutorialId,
-  onEdit,
-  onDelete,
-  onChangeTutorial,
-  ...props
-}: StudentListRowProps): JSX.Element {
-  const classes = useStyles();
-  const additionalMenuItems: ListItem[] = useMemo(() => {
-    const items: ListItem[] = [
-      {
-        primary: 'E-Mail',
-        Icon: MailIcon,
-        disabled: !student.email,
-        onClick: () => {
-          window.location.href = `mailto:${student.email}`;
-        },
-      },
-    ];
-
-    if (!!onChangeTutorial) {
-      items.push({
-        primary: 'Tutorium wechseln',
-        onClick: () => onChangeTutorial(student),
-        Icon: ChangeTutorialIcon,
-      });
-    }
-
-    return items;
-  }, [onChangeTutorial, student]);
-
-  return (
-    <Paper {...props} className={classes.studentBar}>
-      <StudentAvatar student={student} />
-
-      <Box marginLeft={2} minWidth={250} display='flex' flexDirection='column'>
-        <Typography>{student.name}</Typography>
-
-        <Typography variant='body2' color='textSecondary'>
-          {getSubtext({ student, subTextType })}
-        </Typography>
-      </Box>
-
-      <ScheinStatusBox scheinStatus={scheinStatus} marginLeft={2} marginRight='auto' />
-
-      <Button
-        variant='outlined'
-        component={Link}
-        to={ROUTES.STUDENT_INFO.create({
-          tutorialId,
-          studentId: student.id,
-        })}
-        startIcon={<InfoIcon />}
-        style={{ marginRight: 16 }}
-      >
-        Informationen
-      </Button>
-
-      <EntityListItemMenu
-        onEditClicked={() => onEdit(student)}
-        onDeleteClicked={() => onDelete(student)}
-        additionalItems={additionalMenuItems}
-      />
-    </Paper>
-  );
-}
 
 function StudentList({
   students,
@@ -360,37 +157,6 @@ function StudentList({
                   height: elHeight - GUTTER_SIZE,
                 }}
               />
-              //   <StudentAvatar student={student} />
-
-              //   <Box marginLeft={2} minWidth={250} display='flex' flexDirection='column'>
-              //     <Typography>{student.name}</Typography>
-
-              //     <Typography variant='body2' color='textSecondary'>
-              //       {getSubtext({ student, subTextType: studentSubtextType })}
-              //     </Typography>
-              //   </Box>
-
-              //   <ScheinStatusBox scheinStatus={scheinStatus} marginLeft={2} marginRight='auto' />
-
-              //   <Button
-              //     variant='outlined'
-              //     component={Link}
-              //     to={ROUTES.STUDENT_INFO.create({
-              //       tutorialId,
-              //       studentId: student.id,
-              //     })}
-              //     startIcon={<InfoIcon />}
-              //     style={{ marginRight: 16 }}
-              //   >
-              //     Informationen
-              //   </Button>
-
-              //   <EntityListItemMenu
-              //     onEditClicked={() => handleStudentEdit(student)}
-              //     onDeleteClicked={() => handleStudentDelete(student)}
-              //     // additionalItems={additionalMenuItems}
-              //   />
-              // </Paper>
             );
           }}
         </FixedSizeList>
